+++ src/cli/cli-commands/test-command.ts
@@ -44,7 +44,7 @@ export function initializeTestCommand(program: Program) {
             isCacheReloaded,
         }; /* <- TODO: ` satisfies PrepareAndScrapeOptions` */
         const fs = $provideFilesystemForNode(prepareAndScrapeOptions);
-        const llm = $provideLlmToolsForWizzardOrCli(prepareAndScrapeOptions);
+        const llm = await $provideLlmToolsForWizzardOrCli(prepareAndScrapeOptions);
         const executables = await $provideExecutablesForNode(prepareAndScrapeOptions);
         const tools = {
             llm,