+++ src/scrapers/markdown/playground/markdown-scraper-playground.ts
@@ -33,11 +33,11 @@ async function playground() {
     const example = '10-simple.md';
     //               <- TODO: [👩🏿‍🤝‍👩🏼] Read here the examples directory and itterate through all of them
 
-    const llmTools = $provideLlmToolsForTestingAndScriptsAndPlayground({ isCacheReloaded: true });
+    const llmTools = await $provideLlmToolsForTestingAndScriptsAndPlayground({ isCacheReloaded: true });
     const rootDirname = join(__dirname, '..', 'examples');
 
     const markdownScraper = new MarkdownScraper(
-        { llm: $provideLlmToolsForTestingAndScriptsAndPlayground() },
+        { llm: await $provideLlmToolsForTestingAndScriptsAndPlayground() },
         {
             rootDirname,
         },