+++ src/wizzard/wizzard.ts
@@ -73,7 +73,7 @@ class Wizzard {
             isCacheReloaded: false, // <- TODO: Allow to pass
         }; /* <- TODO: ` satisfies PrepareAndScrapeOptions` */
         const fs = $provideFilesystemForNode(prepareAndScrapeOptions);
-        const llm = $provideLlmToolsForWizzardOrCli(prepareAndScrapeOptions);
+        const llm = await $provideLlmToolsForWizzardOrCli(prepareAndScrapeOptions);
         const executables = await $provideExecutablesForNode(prepareAndScrapeOptions);
         const tools = {
             llm,